+++ public/index.php
@@ -4,4 +4,4 @@ use Framework\Kernel;
 require_once(__DIR__ . '/../vendor/autoload.php');
 
 $kernel = new Kernel();
-$kernel->handle($_SERVER['REQUEST_METHOD'], $_SERVER['REQUEST_URI']);
+$kernel->handle($_SERVER['REQUEST_METHOD'], explode('?', $_SERVER['REQUEST_URI'])[0]);