+++ var/cache/templates/4e/4ef8c3755163e818e637da91bfc8870e97ad40479103444dfda787b3eb52186e.php
@@ -1,0 +1,107 @@
+<?php
+
+use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Extension\SandboxExtension;
+use Twig\Markup;
+use Twig\Sandbox\SecurityError;
+use Twig\Sandbox\SecurityNotAllowedTagError;
+use Twig\Sandbox\SecurityNotAllowedFilterError;
+use Twig\Sandbox\SecurityNotAllowedFunctionError;
+use Twig\Source;
+use Twig\Template;
+
+/* layout.html.twig */
+class __TwigTemplate_2c7811e8638b024721db1092e9dd57cdbb8ee8aede2e652dc61e6112c3473d6a extends Template
+{
+    private $source;
+    private $macros = [];
+
+    public function __construct(Environment $env)
+    {
+        parent::__construct($env);
+
+        $this->source = $this->getSourceContext();
+
+        $this->parent = false;
+
+        $this->blocks = [
+            'css' => [$this, 'block_css'],
+            'js' => [$this, 'block_js'],
+            'body' => [$this, 'block_body'],
+            'bottom_js' => [$this, 'block_bottom_js'],
+        ];
+    }
+
+    protected function doDisplay(array $context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        // line 1
+        echo "<!DOCTYPE html>
+<html lang=\"fr\">
+<head>
+    <meta charset=\"UTF-8\">
+    <meta http-equiv=\"X-UA-Compatible\" content=\"IE=edge\">
+    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">
+    <title>Document</title>
+    <link rel=\"stylesheet\" href=\"assets/css/app.css\" />
+    ";
+        // line 9
+        $this->displayBlock('css', $context, $blocks);
+        // line 10
+        echo "    ";
+        $this->displayBlock('js', $context, $blocks);
+        // line 11
+        echo "</head>
+<body>
+    ";
+        // line 13
+        $this->displayBlock('body', $context, $blocks);
+        // line 14
+        echo "    ";
+        $this->displayBlock('bottom_js', $context, $blocks);
+        // line 15
+        echo "</body>
+</html>";
+    }
+
+    // line 9
+    public function block_css($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+    }
+
+    // line 10
+    public function block_js($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+    }
+
+    // line 13
+    public function block_body($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+    }
+
+    // line 14
+    public function block_bottom_js($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+    }
+
+    public function getTemplateName()
+    {
+        return "layout.html.twig";
+    }
+
+    public function getDebugInfo()
+    {
+        return array (  88 => 14,  82 => 13,  76 => 10,  70 => 9,  65 => 15,  62 => 14,  60 => 13,  56 => 11,  53 => 10,  51 => 9,  41 => 1,);
+    }
+
+    public function getSourceContext()
+    {
+        return new Source("", "layout.html.twig", "C:\\wamp64\\www\\MyProject\\board-game\\templates\\layout.html.twig");
+    }
+}