+++ var/cache/templates/8c/8c448274ee355b4e6e182e2b4bc8316a1bf429831dc534ac532ead01149e41e5.php
@@ -1,0 +1,140 @@
+<?php
+
+use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Extension\SandboxExtension;
+use Twig\Markup;
+use Twig\Sandbox\SecurityError;
+use Twig\Sandbox\SecurityNotAllowedTagError;
+use Twig\Sandbox\SecurityNotAllowedFilterError;
+use Twig\Sandbox\SecurityNotAllowedFunctionError;
+use Twig\Source;
+use Twig\Template;
+
+/* home.html.twig */
+class __TwigTemplate_f305e61d8b5d8482146a5f00c08fab7391e1ae9f216665f5e94ac74c2e70bfde extends Template
+{
+    private $source;
+    private $macros = [];
+
+    public function __construct(Environment $env)
+    {
+        parent::__construct($env);
+
+        $this->source = $this->getSourceContext();
+
+        $this->blocks = [
+            'title' => [$this, 'block_title'],
+            'body' => [$this, 'block_body'],
+            'bottom_js' => [$this, 'block_bottom_js'],
+        ];
+    }
+
+    protected function doGetParent(array $context)
+    {
+        // line 1
+        return "layout.html.twig";
+    }
+
+    protected function doDisplay(array $context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        $this->parent = $this->loadTemplate("layout.html.twig", "home.html.twig", 1);
+        $this->parent->display($context, array_merge($this->blocks, $blocks));
+    }
+
+    // line 3
+    public function block_title($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        echo " ";
+        $this->displayParentBlock("title", $context, $blocks);
+        echo " - Accueil ";
+    }
+
+    // line 5
+    public function block_body($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        // line 6
+        echo "<h1>Bonjour ";
+        echo twig_escape_filter($this->env, ($context["firstName"] ?? null), "html", null, true);
+        echo "</h1>
+
+<ul>
+        ";
+        // line 9
+        $context['_parent'] = $context;
+        $context['_seq'] = twig_ensure_traversable(range(1, ($context["loopUntil"] ?? null)));
+        foreach ($context['_seq'] as $context["_key"] => $context["i"]) {
+            // line 10
+            echo "                <li>";
+            echo twig_escape_filter($this->env, $context["i"], "html", null, true);
+            echo "</li>
+        ";
+        }
+        $_parent = $context['_parent'];
+        unset($context['_seq'], $context['_iterated'], $context['_key'], $context['i'], $context['_parent'], $context['loop']);
+        $context = array_intersect_key($context, $_parent) + $_parent;
+        // line 12
+        echo "</ul>
+
+<ul>
+        ";
+        // line 15
+        $context['_parent'] = $context;
+        $context['_seq'] = twig_ensure_traversable(($context["users"] ?? null));
+        $context['_iterated'] = false;
+        foreach ($context['_seq'] as $context["_key"] => $context["user"]) {
+            // line 16
+            echo "                <li>";
+            echo twig_escape_filter($this->env, $context["user"], "html", null, true);
+            echo "</li>
+
+        ";
+            $context['_iterated'] = true;
+        }
+        if (!$context['_iterated']) {
+            // line 19
+            echo "                <li>Aucun users</li>
+        ";
+        }
+        $_parent = $context['_parent'];
+        unset($context['_seq'], $context['_iterated'], $context['_key'], $context['user'], $context['_parent'], $context['loop']);
+        $context = array_intersect_key($context, $_parent) + $_parent;
+        // line 21
+        echo "</ul>
+
+";
+    }
+
+    // line 25
+    public function block_bottom_js($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        // line 26
+        echo "       
+";
+    }
+
+    public function getTemplateName()
+    {
+        return "home.html.twig";
+    }
+
+    public function isTraitable()
+    {
+        return false;
+    }
+
+    public function getDebugInfo()
+    {
+        return array (  117 => 26,  113 => 25,  107 => 21,  100 => 19,  91 => 16,  86 => 15,  81 => 12,  72 => 10,  68 => 9,  61 => 6,  57 => 5,  48 => 3,  37 => 1,);
+    }
+
+    public function getSourceContext()
+    {
+        return new Source("", "home.html.twig", "C:\\wamp64\\www\\MyProject\\board-game\\templates\\home.html.twig");
+    }
+}